Simpler, faster, robuster
The main change is that we no longer rely on layout updates for adding
and removing the preferred dimensions and cleaning up the set of pseudo
windows. This had a few problems:
  - When a window was being floated, its preferred dimensions and status
    as a pseudo window would get removed.
  - Checking for new windows would happen on every single layout update.
  - Checking for new windows required looping over the entire window
    stack.

Now, we no longer listen to layout updates at all and just wait for
window mapping/destroying events to add/remove preferred dimensions and
clean up the set of pseudo windows.

There's also less code! \o/